Huffman_Decode :-> Modified GetMilliSec() time func

+++ Algorithm/HuffMan/include/Huffman_Header.h
@@ -129,7 +129,7 @@ char * GetBinary (uint64_t , int , char *);
 int Percentage_FillUp( int, int , int );
 bool WriteDataIntoFile( int, uint8_t *, int );
 bool GetBitVal( uint64_t, uint8_t );
-long long GetEpochTimeMs(  );
+double GetEpochTimeMs(  );
 
 void BitFeed( uint8_t *, int , bool );
 void BitClear( uint8_t *, int, int );
